resturcture folder and delete unused file
