--- client/public/assets/css/demo.css
@@ -1,124 +1,0 @@
-/*
-* demo.css
-* File include item demo only specific css only
-******************************************************************************/
-
-.menu .app-brand.demo {
-  height: 64px;
-  margin-top: 12px;
-}
-
-.app-brand-logo.demo svg {
-  width: 22px;
-  height: 38px;
-}
-
-.app-brand-text.demo {
-  font-size: 1.75rem;
-  letter-spacing: -0.5px;
-  text-transform: lowercase;
-}
-
-/* ! For .layout-navbar-fixed added fix padding top to .layout-page */
-/* Detached navbar */
-.layout-navbar-fixed .layout-wrapper:not(.layout-horizontal):not(.layout-without-menu) .layout-page {
-  padding-top: 76px !important;
-}
-/* Default navbar */
-.layout-navbar-fixed .layout-wrapper:not(.layout-without-menu) .layout-page {
-  padding-top: 64px !important;
-}
-.docs-page .layout-navbar-fixed.layout-wrapper:not(.layout-without-menu) .layout-page,
-.docs-page .layout-menu-fixed.layout-wrapper:not(.layout-without-menu) .layout-page {
-  padding-top: 62px !important;
-}
-
-/* Navbar page z-index issue solution */
-.content-wrapper .navbar {
-  z-index: auto;
-}
-
-/*
-* Content
-******************************************************************************/
-
-.demo-blocks > * {
-  display: block !important;
-}
-
-.demo-inline-spacing > * {
-  margin: 1rem 0.375rem 0 0 !important;
-}
-
-/* ? .demo-vertical-spacing class is used to have vertical margins between elements. To remove margin-top from the first-child, use .demo-only-element class with .demo-vertical-spacing class. For example, we have used this class in forms-input-groups.html file. */
-.demo-vertical-spacing > * {
-  margin-top: 1rem !important;
-  margin-bottom: 0 !important;
-}
-.demo-vertical-spacing.demo-only-element > :first-child {
-  margin-top: 0 !important;
-}
-
-.demo-vertical-spacing-lg > * {
-  margin-top: 1.875rem !important;
-  margin-bottom: 0 !important;
-}
-.demo-vertical-spacing-lg.demo-only-element > :first-child {
-  margin-top: 0 !important;
-}
-
-.demo-vertical-spacing-xl > * {
-  margin-top: 5rem !important;
-  margin-bottom: 0 !important;
-}
-.demo-vertical-spacing-xl.demo-only-element > :first-child {
-  margin-top: 0 !important;
-}
-
-.rtl-only {
-  display: none !important;
-  text-align: left !important;
-  direction: ltr !important;
-}
-
-[dir='rtl'] .rtl-only {
-  display: block !important;
-}
-
-/* Dropdown buttons going out of small screens */
-@media (max-width: 576px) {
-  #dropdown-variation-demo .btn-group .text-truncate {
-    width: 231px;
-    position: relative;
-  }
-  #dropdown-variation-demo .btn-group .text-truncate::after {
-    position: absolute;
-    top: 45%;
-    right: 0.65rem;
-  }
-}
-
-/*
-* Layout demo
-******************************************************************************/
-
-.layout-demo-wrapper {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-  -ms-flex-direction: column;
-  flex-direction: column;
-  margin-top: 1rem;
-}
-.layout-demo-placeholder img {
-  width: 900px;
-}
-.layout-demo-info {
-  text-align: center;
-  margin-top: 1rem;
-}
