--- client/public/js/bootstrap.js
@@ -1,7 +1,0 @@
-import * as bootstrap from 'bootstrap'
-
-try {
-  window.bootstrap = bootstrap
-} catch (e) {}
-
-export { bootstrap }
