+++ client/src/pages/auth/RegisterPage.jsx
@@ -1,6 +1,5 @@
 import { useState } from "react";
 import { Link } from "react-router-dom"
-import './page-auth.css'
 import { AuthWrapper } from "./AuthWrapper";
 
 export const RegisterPage = () => {
