+++ client/src/pages/errors/ErrorPage.jsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
 import './page-misc.css'
+
 export const ErrorPage = () => {
     return (
         <>
