+++ client/src/pages/layouts/Blank.jsx
