--- client/src/pages/layouts/BlankPage.jsx
@@ -1,8 +1,0 @@
-import React from 'react'
-
-export const BlankPage = () => {
-
-    return (
-        <h4 className="p-4">Blank Page</h4>
-    )
-}
