--- client/src/pages/layouts/FluidPage.jsx
@@ -1,12 +1,0 @@
-import LayoutExample from "../../components/layouts-example/LayoutWrapper";
-
-export const FluidPage = () => {
-  return (
-    <LayoutExample
-      img="/assets/img/layouts/layout-fluid-light.png"
-      alt="Layout fluid"
-      title="Layout Fluid"
-      explanation="Fluid layout sets a <code>100% width</code> at each responsive breakpoint."
-    ></LayoutExample>
-  );
-};
