+++ client/src/pages/layouts/Footer.jsx
@@ -2,13 +2,14 @@ const Footer = () => {
     return (
         <footer className="content-footer footer bg-footer-theme">
             <div className="container-xxl d-flex flex-wrap justify-content-between py-2 flex-md-row flex-column">
-              <div className="mb-2 mb-md-0">
+                <div className="mb-2 mb-md-0">
                 ©
-                  {(new Date().getFullYear())}
+                {(new Date().getFullYear())}
                 , made with ❤️ by Ternakku
-              </div>
+                </div>
             </div>
-          </footer>
-      );
+        </footer>
+    );
 }
+
 export default Footer;
