+++ client/src/pages/layouts/Sidebar.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { Link, NavLink, useLocation } from 'react-router-dom';
-import menuData from '../data/menuData.json'
+import menuData from '../../data/menuData.json'
 
 const Sidebar = () => {
     return (
