--- client/src/pages/layouts/WithoutMenuPage.jsx
@@ -1,11 +1,0 @@
-import LayoutExample from "../../components/layouts-example/LayoutWrapper";
-
-export const WithoutMenuPage = () => {
-  return (
-    <LayoutExample
-      img="/assets/img/layouts/layout-without-menu-light.png"
-      alt="Layout without menu"
-      title="Layout without Menu (Navigation)"
-    ></LayoutExample>
-  );
-};
