--- client/src/pages/layouts/WithoutNavbarPage.jsx
@@ -1,12 +1,0 @@
-import LayoutExample from "../../components/layouts-example/LayoutWrapper";
-
-export const WithoutNavbarPage = () => {
-  return (
-    <LayoutExample
-      img="/assets/img/layouts/layout-without-navbar-light.png"
-      alt="Layout without navbar"
-      title="Layout without Navbar"
-      explanation="Layout does not contain Navbar component."
-    />
-  );
-};
